refactor for PR #1420
